+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 * Fixed bug that shows unfollowers IDs in stdout
 * Fixed bug that made crash the software at the first run
+* Fixed issue: saving after showing unfollowers names, now before
 
 0.4b
 ---